pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.backends"

[project]
name = "mindarchive"
version = "1.1.0"
description = "MindArchive Production Hub — automated faceless YouTube video production"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Kai"},
]
dependencies = [
    # Core
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "typer[all]>=0.12.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",

    # Templates & Frontend
    "jinja2>=3.1.0",
    "python-multipart>=0.0.9",

    # AI & APIs
    "anthropic>=0.45.0",
    "openai>=1.14.0",
    "elevenlabs>=1.2.0",
    "httpx>=0.27.0",

    # Media Processing
    "Pillow>=10.2.0",

    # Storage & Security
    "cryptography>=42.0.0",
    "google-api-python-client>=2.120.0",
    "google-auth-oauthlib>=1.2.0",

    # Configuration
    "tomli>=2.0.0",
    "tomli-w>=1.0.0",

    # Utilities
    "rich>=13.7.0",
    "python-slugify>=8.0.0",
    "aiofiles>=23.2.0",
    "sse-starlette>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]

[project.scripts]
mindarchive = "mindarchive.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/mindarchive"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "RUF"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
